# src/api/data/swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: Hammerbarn API
  description: >-
    Provides a comprehensive listing of products available for sale, including
    detailed product information and attributes
host: localhost:3000
basePath: /
schemes:
  - https
consumes: []
produces:
  - application/json
paths:
  /api/listings/search:
    get:
      summary: Get a list of products matching the specified criteria
      description: >-
        Returns a list of products, optionally filtered by providing a
        description of your needs
      operationId: GetProductListings
      parameters:
        - name: query
          in: query
          type: string
          required: true
          description: Your product search query
          x-ms-summary: Product Query
          x-ms-visibility: important
      responses:
        default:
          description: List of products matching the specified criteria
          schema:
            type: array
            items:
              type: object
              properties:
                uid:
                  type: string
                  description: Unique identifier of the product
                  title: Id
                  x-ms-visibility: important
                name:
                  type: string
                  description: Name of the product
                  title: Name
                  x-ms-visibility: important
                description:
                  type: string
                  description: A detailed description of the product
                  title: Description
                  x-ms-visibility: important
                attributes:
                  type: array
                  items:
                    type: object
                    properties:
                      key:
                        type: string
                        description: The attribute type
                        title: Key
                        x-ms-visibility: important
                      value:
                        type: string
                        description: The attribute value
                        title: Value
                        x-ms-visibility: important
      x-ms-openai-data:
        openai-enabled: true
        operations:
          - operationId: GetProductListings
            x-ms-require-user-confirmation: false
            summary: Get a list of products matching the specified criteria
            description: >-
              Returns a list of products, optionally filtered by providing the
              product name, description, or an array of key-value attributes as
              query parameters
      x-ms-visibility: important
  /api/listings/product/{id}:
    get:
      summary: Get an individual product by ID
      description: >-
        Returns the details of a specific product by its unique identifier
      operationId: GetProductById
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: The unique identifier of the product
      responses:
        200:
          description: Product details retrieved successfully
          schema:
            type: object
            properties:
              uid:
                type: string
                description: Unique identifier of the product
              name:
                type: string
                description: Name of the product
              description:
                type: string
                description: A detailed description of the product
              attributes:
                type: array
                items:
                  type: object
                  properties:
                    key:
                      type: string
                      description: The attribute type
                    value:
                      type: string
                      description: The attribute value
        default:
          description: Error retrieving product details
          schema:
            $ref: '#/definitions/ErrorResponse'
definitions:
  ErrorResponse:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
parameters: {}
responses: {}
securityDefinitions: {}
security: []
tags: []
x-ms-openai-manifest:
  humanName: Hammerbarn
  modelDescription: >-
    Plugin for accessing a wide range of products for sale. Use it whenever a
    user asks about product listings available in the retail market. This plugin
    can be used to search for products by name, description, or by specific
    attributes.
  contactEmail: contact@retail.hammerbarn.com
  legalUrl: https://hammerbarn.com/legal/